[CLI] Assume upload if first arg is a valid file
Since the CLI is being designed with simplicity in mind, a lot of
actions should be implicit. If a valid file path is passed to yeetfile,
the user is trying to upload a file. If a URL is passed to yeetfile (or
in the future, a valid URL path string like "this.blah.thing"), they are
trying to download a file.