[CLI] Support upload+download files of any size
Upload and download is now functional using the CLI for files of any
size. At some point this should be updated to make sure a user has a
paid account before allowing uploads of a certain size.

A `shared` folder was added to encapsulate anything that would make
sense for the CLI and server to share between one another.