[DB] Track B2 id and output path in metadata table
The ID for a file in B2 is necessary for download and needs to be
tracked by Yeetfile after it has been successfully uploaded.

Output path (the path used for accessing the file later) is also needed
in order to perform a lookup and actually find the file metadata for
downloading.