Use gorilla/sessions for session mgmt, update router
gorilla/sessions will be used for handling individual user sessions,
although something else will likely need to be used (API key?) for the
command line interface of yeetfile.

The router was updated to use a new Route struct as the key in the
router map (previously just a string path). This allows explicitly
stating the http method that must be used for a particular handler, and
allows separating GET functionality for a route (i.e. returning HTML for
GET requests to /signup) from POST functionality (i.e. signing a user up
witha  POST request to /signup).