[CLI] Encrypt salt w/ separate key before upload
As an added precaution, the salt for the primary key that is used for
encrypting chunks of file data is encrypted with a separate key before
it is uploaded to the server. In the case of a compromised databse, an
attacker would not have access to unencrypted salts, which could be used
in an attempt to brute force a file's password.

This is implemented in the CLI app, but not in the web yet.