Remove `path` from metadata table, use pepper w/ pw
The `path` field in the metadata table isn't really useful, since it
just adds an extra step to access the file. This was initially done in
order to allow someone to easily communicate the path to the file
verbally (if necessary). Instead, the file ID is used to access the
file, but the file's key is generated with a "pepper" appended to the
password.

This new pepper implementation replaces the previous change that
encrypted the salt before sending it to the server.