Accept payments via Stripe
This handles the basic integration steps for accepting payment for
YeetFile products (meter increases) through Stripe.

Payments made through Stripe have a webhook event that hits the /stripe
endpoint, which handles checkout.session.completed by updating the
user's meter and storing the transaction in the database. This last part
should be updated to be temporary and deleted after a certain amount of
time has passed, since it's only useful for resolving immediate issues
with payment.