[CLI] Use app dir for config + session data
The app now uses a unique directory ($HOME/.config/yeetfile) for storing
items such as a config.yml (which can be used to modify the server that
gets used by the command line tool) and a session file. A gitignore is
also provided to ignore the session file, in case a user is tracking
their .config directory in git.