[DB] Fix user ID generation
User IDs were previously being generated in a way where a []byte of
random numbers was converted to a string, which caused user IDs to be a
sequence of characters instead of numbers. This new approach uses an
array of runes instead of an array of bytes, which produces the expected
output.