Finalize login/logout routes
Switched sessions to filesystem store to allow persisting between
reboots

Added CLI commands for logging in/out and prompting for login if the
user doesn't have a valid session when uploading. This provides a better
UX than starting the upload and then being rejected.

Added /session route to quickly check if the user has a valid session.
This is really only useful for the CLI app, since the web interface will
just redirect to the login page.