Implement email verification on signup
Users signing up with their email address will now have their email
address verified by having a six digit code sent to their email. By
either clicking the link or entering the code on the page (or in their
console), their account is "verified" and moved to the users db table.

This adds a new "verify" table that contains an email, verification
code, password hash, and date. The email and pw hash are used when
verifying the user and creating a new user entry, and the date
can/should be used for checking to see how long it's been since the last
verification code was sent. There should be a time limit for when the
user can re-request a verification code, as well as an expiration for a
point when the verification entry can be assumed abandoned and can be
deleted.