Update DB conn, gen init sql from new script
The initial state of the db is now created from a script that can use
environment variables as needed.

The db conn string was updated to use a url style string -- for some
reason, changing the user to "yeetfile" caused the previous conn string
to always try connecting to a "yeetfile" database, even though dbname
was clearly set to "postgres". I didn't care enough to look into this
further since the url style string works fine.