Require membership activation for file upload
Rather than trying to figure out a way to prevent user's abusing a free
tier, there's now a requirement to purchase either a 1-month or 1-year
membership, which can be further modified by transfer limit upgrades as
needed.

This could be changed down the road to set new users up with a minimal
amount of file transfer to try it out, but for the beta it seems to make
sense to only allow paid (or users with a promo code) to use the
service.

Memberships are set to $2 for 1 month, or $9 for 1 year. Neither
membership tier auto-renews.