Use AES-GCM from Web Crypto instead of libsodium
Libsodium crypto required third party javascript libraries to be
imported, which isn't bad, but it's a better idea to just use the built
in key derivation + en/decrypting via the browser's web crypto API.

This required refactoring the CLI as well to use AES-GCM in place of
secretbox/libsodium.

Functionally everything works the same.