Add distinction between "sending" and "storage" keys
YeetFile is now going to move in the direction of supporting storing
files instead of just sending them, since it isn't too much extra work
and a lot of the existing cryptography on the clients (and the backend
logic as well) is built to support this.

The keys will be treated differently though. There will need to be a
"master" key for file storage that is generated and encrypted client
side with the user's password, and that key should be encrypted before
being sent to the server. The server can then return this key on login,
and the client can decrypt the key (and then use that key to decrypt
files sent back).