Implement key system for storing files
To add more overall value to YeetFile, the ability to store files
(rather than only being able to send them) is being added. This involves
a new key system that will allow securely storing encrypted files and
decrypting them at a later date without having the server ever see the
user's password or primary/user key.

The new system involves:

- A user key (aka primary, master, etc)
- A login key
- A storage key

The user key is a PBKDF2 key generated from the user's password as the
payload and their identifier (either email or account ID) as the salt.

The login key is generated from the user key as the payload and the
user's password as the salt.

The storage key is a 32-byte CSPRNG generated value that is encrypted
with the user's user key before being sent to the server on signup. The
storage key will then be returned on login and used client-side to
decrypt all stored files.

This change involved adding passwords for account ID only signup.