Cache downloaded files to disk
This adds support for caching files to the disk rather than having each
download pass through B2. The size of the cache directory is
configurable using environment variables to specify the max total size
of the directory as well as the max file size to be inserted into the
cache.

The cache removes files based on the oldest file (either by creation/mod
or when it was last downloaded) as long as the file isn't currently
being downloaded.