Implement file vault/storage
This adds a new "vault" feature to YeetFile, which allows for uploading
files for long-term storage rather than just the original ephemeral file
sharing functionality of YeetFile.

There are a number of cryptography changes that were made as a part of
this feature:

- Users now have a key pair, which is used for file/folder
  encryption/decryption. The user's public key is stored in the users
  table and can be used by other users when sharing a file or folder
  with that user. The user's private key is encrypted using their PBKDF2
  generated "user key" described next.
- Users now have a "user key" which is never shared with the server,
  but is used to encrypt the user's private key, and for generating the
  user's "login key", described next.
- Users now have a "login key", which is the user key but passed through
  PBKDF2 again with the user key as the payload and the user's password
  as the salt.

Each user has a default root folder in their vault that has the same ID
as their user ID (which is never displayed to the user). Files and
folders created in the vault each have their own randomly generated key,
which is encrypted in two different ways. If the file/folder is uploaded
to the root folder, this key is encrypted using the user's public key.
If the file/folder is uploaded to a subfolder, the key is encrypted
using the subfolder's key. When accessing a subfolder, a key sequence
needs to be retrieved from the database containing the keys -- in order
-- to decrypt in order to read the contents of the folder.

For example, given /root/folderA/folderB structure, if we want to access
folderB, we would need to retrieve the root folder key and the folderA
key. The root folder key would be decrypted using the user's private
key, and the folderA key would be decrypted using the decrypted root
folder key. Then the decrypted folderA key could be used to decrypt
folderB.