Improve server configuration options
This implements a new configuration approach, where config settings are
parsed into structs when the service is initialized, and can inform
feature availability throughout the rest of the app.

For example, a BTCPayConfig struct is now parsed on init, and if any
field is missing or invalid, BTCPay would not be available as a billing
option on checkout.

The overall configuration state of the server is displayed in the
console when YeetFile is launched.