Update billing to match new subscription model
Stripe and BTCPay payments have been updated to reflect how
subscriptions work in YeetFile now.

Stripe payments auto-renew and can be managed/upgraded/downgraded
through the Stripe customer portal. The link for the portal is generated
by using the "manage-sub" endpoint with the user's payment ID.

BTCPay payments do not auto-renew, and cannot be
managed/upgraded/downgraded/etc. BTCPay payments use the "Forms"
feature, with a unique form for each subscription. This replaces the
previous approach of using the Invoices API to generate invoices for
each customer to pay. This simplifies a lot of the logic for receiving
payments from BTCPay.