Implement full CLI/TUI for YeetFile
This introduces a full-featured CLI/TUI for YeetFile that supports the
same core features as the web app, namely:

- Signup
- Login/Logout
- Vault storage
- Sharing files/folders in vault
- Ephemeral sending of files or text

Account management is not implemented yet.

This also irons out a number of issues on the backend, such as
inconsistency of response types for similar API requests, using shared
endpoint strings to define routes, improper handling of shared folder
access, and more.