Refactor all JS to TS, add Go -> TS -> JS compiling
This removes all JS (except for JSZip and StreamSaver dependencies) and
replaces it with TS. Some TS is still plain JS, but will be improved
gradually over time. The unchanged code is being kept the same because
it's likely to change with the next few branches.

This introduces a way of compiling Go structs to TS interfaces, which
allows much better control over how the web client interacts with the
YeetFile API.