Implement automated testing of Go and TS
This adds github actions to test Go and TypeScript code. Currently the
Go code tests cryptography related functions, some utilities, and the
majority of api interactions. The TypeScript tests only test the
cryptography functions. API tests are not as crucial for the TS tests
since they should be functionally identical to the Go API tests.