Allow server-specific password to restrict signup
YEETFILE_SERVER_PASSWORD can now be set when running yeetfile-server to
restrict signups only to those who know the server password. This
password is ingested as a bcrypt hash when the server starts and is
required to be included in the initial signup request before the signup
process can proceed.