Allow deleting accounts, add CLI account view
This implements the ability for users to delete their own accounts. The
process on CLI and Web is to navigate to the Accounts view, then select
"Delete Account". The user will be prompted to enter their login info in
order to finalize deleting their account.

This also adds the account view for CLI, accessible via "yeetfile
account".