[Web] Implement simultaneous chunk uploads
Similar to the CLI, the web now uploads multiple chunks at once.

The multi-file upload has also been disabled for now, since it seems to
cause OOM issues on browsers when multiple large (>1GB) files are
selected (but not with individual files that exceed the size of multiple
files combined).