Add ability to change password
When a user changes their password, they're required to enter their
current login credentials along with their desired new password. The
current login credentials are used to recreate their user key, which can
be used to decrypt the protected private key and re-encrypt it with the
new user key generated from the new password. The new login info is also
used to generate a new login key hash, which is sent to the server and
bcrypt hashed before updating in the database.