Rework email verification flow
Email verification endpoints are now split between the HTML page
(/verify_email) and the handler for actually verifying the user's email
(/api/.../verify_email).

The verify table now has an expiration date included of 5 minutes past
the point of the initial request. If another request comes in before
that 5 minutes has elapsed, the verification values are updated but a
new code is not generated (and not sent to the user).