Implement auth based limiter + blake2 hash limiter id
This introduces an auth based limiter, which prevents authenticated
users from making the same request multiple times in a row in a short
period of time. This is mainly only used to prevent intentionally
requesting the same file chunk multiple times in a row.

Limiter management has been added to the cron tasks on startup.

Account IDs and IP addresses for both limiters are now hashed via blake2
before storing for added security.