Combine server + client err msgs in debug mode
This introduces a function that produces a server and client error
message, where the former contains specific details regarding the error
(which could possibly contain information the client shouldn't receive).
In debug mode, this server specific error information is included in the
client message as well in order to aid with debugging.