Add daily bandwidth "soft cap"
This introduces a daily bandwidth limitation of 3x a user's available
storage. Once this bandwidth limit has been reached for the day, the
user will be locked out of downloading files until the next day. This is
described as a "soft cap" because this can be removed by emailing YF
support (if on the official instance).

The limitation on bandwidth is meant as a way to deter potential script
kiddie activity where a file is continuously downloaded in an attempt to
incur large egress fees.