Implement password hint system
This introduces a password hint system for email-based users of
YeetFile. On signup, or at any point after creating an account, users
can set a password hint that is sent to the server. This is sent
unencrypted, but is encrypted by the server before storing in the
database using the YEETFILE_SERVER_SECRET variable. This has to be
something encrypted/decrypted by the server so that a user can request
their password hint to be emailed to them if they ever forget their
password.

The login view on web and CLI now includes a "forgot password" option
that allows the user to request their password hint to be emailed to
them. If the specified email doesn't exist in the user table, or if the
user doesn't have a password hint enabled, then nothing happens.

If the user has set a password hint but wants to remove it from their
account, they can use the "change password hint" option in their account
view and set the password hint to an empty value.

Requests for the password hint must be at least 1 hour apart, and are
currently not removed from the database by any existing process.