Add ability to change or set email
Email users can now change their email. This involves sending a
link to the user's existing email, which provides them with a
unique ID that must be used to update their email address. They provide
their login info and new email, and a verification code is sent to the
new email to confirm.

Account ID only users can also set up an email. The process is identical
to changing an email, except it skips the very first step of sending the
unique link to the user's current email.

The `verify` table was updated to include a column for account ID, which
can be used to indicate when an email verification should update an
existing account rather than create a new account.

Other various small changes:
- `YEETFILE_CALLBACK_DOMAIN` -> `YEETFILE_DOMAIN` (since it's used for
  more than just the email callback
- Added default security headers for all requests
- Renamed `upload.css` -> `send.css` since that's more accurate