Implement two-factor auth
This adds support for TOTP based 2FA for both web and CLI.

The `users` table now has a `secret` column and a `recovery_hashes`
column. The `secret` column contains the TOTP secret and is encrypted
using the server-side key. The `recovery_hashes` column contains 6
bcrypt hashes of 6 random codes that are sent to the user. If a user
uses one of these 6 codes instead of their 2FA code, the hash is removed
from the database.

This also includes the ability to remove 2FA after enabling it, which
requires the 6-digit code or a recovery code to disable.