Add vault file preview for web and CLI
This introduces the ability to preview a file directly from within a
vault folder (depending on file type).

On both clients, this involves checking if the file is under a certain
size limit (< 5 chunks) and if so, downloads and decrypts the file into
memory. On web, the decrypted file is used to create a blob URL that
allows easily embedding the decrypted content into the vault view. On
CLI, the decrypted file is kept in memory and passed to a viewer method
that can display some files, but not all.

In the CLI app, the types of files that can be previewed is much smaller
for obvious reasons. Text-only files can be previewed without any issue,
and certain image types (png or jpg) can be translated to ascii and
displayed in the view. This can potentially be expanded upon at some
point, but is out of scope for the foreseeable future.