[Web] Refactor indexeddb enc key gen
Previously, the key pair inserted into indexeddb was encrypted using a
random value provided by the server. This was prone to issues with users
who might not have caching enabled in their browser, for security
reasons or otherwise.

This approach has been refactored to instead return a blake2 sha256 hash
of the user's session key + session id. This way the indexeddb has a
session-specific key to encrypt the user's key pair that is immune to
issues with browser caching.

This requires dynamically importing db.js after authentication, which
meant updating from es2017 to es2020.