Implement custom filepicker, finish CLI acct controls
The filepicker native to bubbletea doesn't allow filtering within the
current folder, which makes it somewhat useless in directories with a
lot of files and folders. This adds a custom filepicker built on top of
the bubbletea List model, which does have support for filtering by a
struct's property (in this case, file or folder name).

A new endpoint has been added for fetching info about how the current
server has been configured, which can then be used for things like
determining if email signup is possible, checking if billing is set up,
etc.

CLI account controls have been finished up, allowing users to initiate
checkout from the CLI and not be required to log into the web interface.