Implement password vault
This splits the existing vault feature to allow storing files (same as
before) as well as passwords/logins/etc.

The initial vault upload struct has been updated to also include a field
for encrypted byte data representing a shared.PassEntry struct that has
been encrypted with the item key. This allows password entries to be
uploaded using a single API call.

Rather than storing the password data in Backblaze, the data is stored
in the postgres db in a new column `pw_data` in the vault table. This
allows faster access of password entries, since everything can be
decrypted directly from the folder API response.

This also includes a password generator, which is split between regular
"password" generation (random characters, numbers, symbols, etc) and
"passphrase" generation. The latter of these involves downloading
wordlist json files containing long and short words that are either
stored in indexeddb on web or in the user's yeetfile config dir when
using the cli app.

All of the same sharing functionality is available for password entries
since they're essentially just vault files with extra metadata.

Ultimately the password vault lacks a lot of features that would be
important for everyday use (no search, no browser extension, etc), but
is intended to be a good foundation that can be built upon moving
forward.