Improve CLI help formatting, refactor vault pkg names
CLI help is now formatted a bit cleaner, with an additional help guide
for filtering results. The main change was making spacing consistent
with the vault view and removing the extra line break before the help
section.

`files` pkg in CLI app has been renamed to `items` to encapsulate that
it is responsible for displaying both files and password entries,
depending on the selected view