[CLI] Store server info requests in user config dir
The server info for a user's configured server is unlikely to change
often, but was being requested on startup of the cli tool each time.

This changes the behavior to only re-request server info from the server
once every 24 hours. This could be extended more in the future,
depending on user feedback.