Validate upload URL before initiating large file upload
There seems to be a situation where a large file upload can be marked as
non-local and attempt to POST file contents to an invalid URL (in this
case, the name of the local upload directory).

This obviously results in an error. This change ensures that the
designated upload URL is actually a valid URL, otherwise it forces the
local-only "dummy" mode to be enabled.

See #14