Add instance admin web UI (#18)
This implements an admin UI for instance maintainers to allow content
moderation as needed.

Basic initial features include:
- User lookup
- Listing files uploaded by a user
- File lookup
- File deletion
- User deletion

Admin declaration is handled with the YEETFILE_INSTANCE_ADMIN environment
variable, which can be set to an admin's email or account ID to enable.

Additional metadata was added to files uploaded via YF Send, since the same
need for moderation applies to those files as well. 

Currently this is web-only since instance maintainers are mostly using the web
interface from what I've gathered, but will still be ported to the CLI app
soon.