Init remote upload checksums w/ placeholder value
If chunks are received out of order (ie chunk 2 before 0), the query to
update and return current checksums breaks down since the pq.Array
parser must find an element in index 0 for the returned value to be
considered a valid array.

This initializes checksums with a "?" placeholder as the first element
in the checksums array, which is overwritten once chunk 0 is handled.